# docs/api/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Telegram Bot Builder API",
    "description": "REST API визуального конструктора Telegram-ботов. Авторизация: сессионная cookie или Bearer PAT. Документация: /admin (вход по ADMIN_API_KEY) → /admin/docs.",
    "version": "2.2.0.5"
  },
...
  "paths": {
...
    "/admin/api/app-settings": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "Настройки приложения (по провайдерам)",
        "description": "Требует admin cookie после `/admin/login`. Секреты и токены **не** возвращаются — только флаги `*Configured`.",
        "security": [
          {
            "adminCookie": []
          }
        ],
        "responses": {
          "200": {
            "description": "Текущие настройки",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminAppSettingsResponse"
                }
              }
            }
          },
          "401": {
            "description": "Не авторизован в admin",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "admin"
        ],
        "summary": "Сохранить настройки приложения",
        "description": "Upsert по секциям провайдеров. Пустой `clientSecret` / `botToken` не удаляет существующие значения. `botUsername` опционально — резолв через getMe при заданном token.",
        "security": [
          {
            "adminCookie": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdminAppSettingsPayload"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Настройки сохранены",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminAppSettingsSave"
                }
              }
            }
          },
          "400": {
            "description": "Валидация",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "telegram.clientId обязателен"
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Не авторизован в admin",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "telegram.clientId обязателен"
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          },
          "500": {
            "description": "Внутренняя ошибка",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "telegram.clientId обязателен"
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "connect.sid",
        "description": "Сессия после успешного login (POST /api/auth/telegram или miniapp/dev-login)"
      },
      "agentToken": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "PAT",
        "description": "Персональный токен агента (MCP/CLI)"
      },
      "adminCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "admin_auth",
        "description": "Admin cookie после POST /admin/api/login (ADMIN_API_KEY)"
      }
    },
    "schemas": {
...
      "AdminTelegramProvider": {
        "type": "object",
        "properties": {
          "clientId": {
            "type": "string",
            "example": "123456789"
          },
          "botUsername": {
            "type": "string",
            "example": "my_bot"
          },
          "clientSecretConfigured": {
            "type": "boolean",
            "example": true
          },
          "botTokenConfigured": {
            "type": "boolean",
            "example": true
          },
          "configured": {
            "type": "boolean",
            "example": true
          }
        },
        "required": [
          "clientId",
          "botUsername",
          "clientSecretConfigured",
          "botTokenConfigured",
          "configured"
        ]
      },
      "AdminAppSettingsResponse": {
        "type": "object",
        "properties": {
          "configured": {
            "type": "boolean"
          },
          "providers": {
            "type": "object",
            "properties": {
              "telegram": {
                "$ref": "#/components/schemas/AdminTelegramProvider"
              }
            },
            "required": [
              "telegram"
            ]
          }
        },
        "required": [
          "configured",
          "providers"
        ]
      },
      "AdminAppSettingsSave": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "enum": [
              true
            ]
          },
          "configured": {
            "type": "boolean"
          },
          "providers": {
            "type": "object",
            "properties": {
              "telegram": {
                "type": "object",
                "properties": {
                  "botUsername": {
                    "type": "string"
                  },
                  "configured": {
                    "type": "boolean"
                  }
                },
                "required": [
                  "configured"
                ]
              }
            },
            "required": [
              "telegram"
            ]
          }
        },
        "required": [
          "success",
          "configured",
          "providers"
        ]
      },
      "AdminTelegramSettingsPayload": {
        "type": "object",
        "properties": {
          "clientId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "number"
              }
            ],
            "example": "123456789"
          },
          "clientSecret": {
            "type": "string"
          },
          "botUsername": {
            "type": "string"
          },
          "botToken": {
            "type": "string"
          }
        },
        "required": [
          "clientId"
        ]
      },
      "AdminAppSettingsPayload": {
        "type": "object",
        "properties": {
          "telegram": {
            "$ref": "#/components/schemas/AdminTelegramSettingsPayload"
          }
        },
        "required": [
          "telegram"
        ]
      },
...
    },
...
  },
  "security": [
    {
      "cookieAuth": []
    },
    {
      "agentToken": []
    }
  ]
}